Move query type mux example to separate test package
